Implement trash feature: files are not deleted but moved to a Trash folder, where they can be restored, or will be deleted after 30 days
FossilOrigin-Name: 3f40dfa9aac5fc7c2506177fd3267981daf7178276e5afead843947844f5da02

+++ src/include/lib/Garradin/Files/WebDAV/Storage.php
@@ -349,7 +349,12 @@ class Storage extends AbstractStorage
 			throw new WebDAV_Exception('Vous n\'avez pas l\'autorisation de supprimer ce fichier', 403);
 		}
 
-		$target->delete();
+		if ($file->context() == $file::CONTEXT_TRASH) {
+			$target->delete();
+		}
+		else {
+			$target->moveToTrash();
+		}
 	}
 
 	protected function copymove(bool $move, string $uri, string $destination): bool
